rename : AllEvent -> page폴더로 옮김
